Usability Testing Adjustments
I kept the keys and letters of the game in the QWERTY format that everyone is familiar with due to keyboards so they can easily tell where letters are. I also added a light yellow effect when clicking the buttons so there is not an out of place blue colour which does not fit the colour scheme. I also added a game icon to the game to give it more appearance and describe what it is. Finally, I added the indicator of what round it is during progressive mode so they can keep track while they make their way up the rounds to the more difficult words. The end page was also adjusted so each component is now more spaced out and not all cluttered in the one spot that it was. The word display was also reduced in size